requirements.txt
selenium
webdriver-manager
requests
python-dotenv
elasticsearch